+++ app/src/main/java/com/sun_asterisk/myeditor03/screen/main/MainActivity.kt
@@ -5,16 +5,15 @@ import android.support.v7.app.AppCompatActivity
 import com.sun_asterisk.myeditor03.R
 import com.sun_asterisk.myeditor03.R.layout
 import com.sun_asterisk.myeditor03.screen.home.HomeFragment
-import com.sun_asterisk.myeditor03.utils.FragmentHelper
+import com.sun_asterisk.myeditor03.utils.goNextChildFragment
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(layout.activity_main)
-        FragmentHelper.goNextChildFragment(
-            supportFragmentManager, R.id.layoutContainer,
-            HomeFragment.instance(), false, HomeFragment.TAG
-        )
+        goNextChildFragment(
+            R.id.layoutContainer,
+            HomeFragment.instance(), false, HomeFragment.TAG)
     }
 }
