+++ app/src/main/res/values/dimens.xml
@@ -1,20 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="dp_4">4dp</dimen>
-    <dimen name="dp_8">8dp</dimen>
-    <dimen name="dp_12">12dp</dimen>
     <dimen name="dp_16">16dp</dimen>
     <dimen name="dp_18">18dp</dimen>
-    <dimen name="dp_24">24dp</dimen>
-    <dimen name="dp_40">40dp</dimen>
-    <dimen name="dp_80">80dp</dimen>
-    <dimen name="dp_100">100dp</dimen>
-    <dimen name="dp_140">140dp</dimen>
-    <dimen name="dp_180">180dp</dimen>
-    <dimen name="dp_200">200dp</dimen>
-    <dimen name="sp_12">12sp</dimen>
-    <dimen name="sp_16">16sp</dimen>
-    <dimen name="sp_18">18sp</dimen>
-    <dimen name="sp_20">20sp</dimen>
-    <dimen name="sp_30">30sp</dimen>
 </resources>
